Moved analysis and evaluation to analysis.py
