+++ project1/src/regression.py
@@ -22,36 +22,6 @@ def FrankeFunction(x,y):
 z = FrankeFunction(x, y)
 
 # Function definitions
-def mean_squared_error(y, y_predicted):
-    """
-    Evaluate the mean squared error of the output generated
-    by Ordinary Least Squares regressions.
-    y : the data on which OLS was performed on
-    y_predicted : the data the model spits out after regression
-    """
-
-    N = y.shape[0]
-    mse = np.sum(np.square(y - y_predicted))/N
-    return mse
-
-def r2_score(y, y_predicted):
-    """
-    Evaluate the R2 score function.
-    y : the data on which OLS was performed on
-    y_predicted : the data the model spits out after regression
-    y_mean : mean value of y
-    upper/lower_sum : numerator/denominator in R2 definition.
-    """
-    N = y.shape[0]
-    y_mean = np.sum(y)/N
-    upper_sum = np.sum(np.square(y - y_predicted))
-    lower_sum = np.sum(np.square(y - y_mean))
-    r2score = 1 - upper_sum/lower_sum
-
-
-# Setting up the matrix X in the matrix equation y = X*Beta + Eps
-# and perform regression
-
 def ordinary_least_squares(y, m):
     """
     Perform linear regression using th ordinary least squares method
@@ -61,6 +31,9 @@ def ordinary_least_squares(y, m):
     and X contains the polynomial expressions.
     returns x and y values for plotting along with the predicted y values
     from the model.
+
+    Sets up the matrix X in the matrix equation y = X*Beta
+    and performs regression
     """
 
     # Setup
@@ -80,32 +53,3 @@ def ordinary_least_squares(y, m):
 
 
 output = ordinary_least_squares(z,5)
-
-# =================================================================
-# Plotting
-# Plots the least squares fit and the FrankeFunction next to
-# each other for comparison.
-fig = plt.figure()
-ax = fig.add_subplot(1, 2, 1, projection='3d')
-
-surf1 = ax.plot_surface(output[0], output[1], output[2], cmap=cm.coolwarm,
-                       linewidth=0, antialiased=False)
-# Customize the z axis.
-ax.set_zlim(-0.10, 1.40)
-ax.zaxis.set_major_locator(LinearLocator(10))
-ax.zaxis.set_major_formatter(FormatStrFormatter('%.02f'))
-
-# Add a color bar which maps values to colors.
-fig.colorbar(surf1, shrink=0.5, aspect=5)
-
-# Second subplot
-ax = fig.add_subplot(1,2,2, projection='3d')
-surf2 = ax.plot_surface(x, y, z, cmap=cm.coolwarm,
-                       linewidth=0, antialiased=False)
-ax.set_zlim(-0.10, 1.40)
-ax.zaxis.set_major_locator(LinearLocator(10))
-ax.zaxis.set_major_formatter(FormatStrFormatter('%.02f'))
-
-fig.colorbar(surf2, shrink=0.5, aspect=5)
-
-plt.show()
